like and dislike profile update

+++ src/controller/comment-controller.js
@@ -1,4 +1,5 @@
 import { deleteComment, updateComment } from '../model/user-post.js';
+import { currentUser } from '../model/auth-user.js';
 
 export const showComments = (event) => {
   const btnShowComment = event.target;
@@ -8,14 +9,20 @@ export const showComments = (event) => {
 export const eventDeleteComment = (event) => {
   const btnDeleteComment = event.target;
   const idComment = btnDeleteComment.closest('.comment').id;
-  deleteComment(idComment)
-    .then((doc) => {
-      console.log('comentario eliminado', doc);
-    })
-    .catch((err) => {
-      console.log(err);
-    });
+  const userId = btnDeleteComment.closest('.card-post').querySelector('.user-post').id;
+  if (currentUser().uid === userId) {
+    deleteComment(idComment)
+      .then((doc) => {
+        console.log('comentario eliminado', doc);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  } else {
+    console.log('usted no puede eliminar este comentario, porque pertenece a otro usuario');
+  }
 };
+
 
 export const eventUpdateComment = (event) => {
   const btnUpdateComment = event.target;
